Fix OpenGL errors for macOS
Force a 2.1 compatability context and bypass the check for 3.1 support. This is hard-coded generally, so a mac-specific fix should be implemented before this becomes official.

Since Mac has no shader support (OpenGL 2.1 compatability is its latest compat context) comment out the shaders to reduce errors in the log. This should be reverted before merging into maintenance, and a more formal solution considered.

Mac’s inbuilt GLU headers seem incompatable with the C version being used, so we’ve redirected to Spring’s. *This has been tested with Spring 103.0.1 and nothing seems to break.* Treat this change with fair skepticism also.