Skip shatter pieces on macOS
A segfault that occurs in the creation of shatter variants that I don’t want to focus on fixing quite yet. Initial research implies it’s hitting a bug inside of an external library, however this conclusion should be treated with skepticism.

An alternative fix should be considered before merging back into the official maintenance branch.