Add Prototype `Renderer` class
Handles rendering for `SpringApp` (Setup + splash screen) and `CPreGame`.

Currently rendering works up to and for the splash screen, and not beyond. Further functionality will not be hard to add.

# Rationale

Graphics/presentation code is currently tied into the engine, where for e.g. Headless code requires numerous compiler directives to isolate behaviour. Separating it out into self-contained code will not only mitigate this need but potentially allow simultaneous development of two rendering pipelines on the same engine version.

# Structure

The “Renderer” is now a standalone object, which wraps the general structure of the calls in `SpringApp.cpp`. `Renderer` itself does not implement rendering, but leaves that to subclasses such as `GLRenderer`. (Headless rendering can be obtained by instantiating `Renderer` directly, or preferrably through its alias `HeadlessRenderer`).

Only `SpringApp` and `CGameController` know about the renderer. For drawing, `CGameController::Draw()` has been re-assigned to `Renderer`, which will forward it to the appropriate `Scene`.

`Renderer` uses the idea of `Scenes` to handle controller-specific rendering. Think of `Scene` as just wrapping`CGameController::Draw()`. Scenes are ususlly going to be created with a reference to the controller they perform draws for.

Note that the API that `Renderer` exposes has not been modified from the preexisting calls in `SpringApp` and  `CGameController`. API revision will occur later.

# What about widgets/UI?

Lua rendering is likely to not easily fit into this paradigm; either widgets could adapt depending on the renderer they’re given, or they could (more reasonably) be provided a renderer-agnostic API, with extensions. Unimplemented functions would simply do nothing.

This API would require a core set of functions that handle essential drawing, such as UI. These functions would be required for any `Renderer` to be considered functional (excepting headless).